remove temp debugging logic
introduced a new environment variable and encountered a maintenance cost.  also,
shouldn't be needed as a breakpoint can be set on the error statement instead of
adding a manual pause.  alternatively, we could dump a stacktrace (with or without
bombing) to catch them or do what libc does and simply ignore them if they are no
longer useful to us.