no longer make bu_free() blather on NULL free, instead matching stdlib's silent handling.
keeping the -1L test as a way for callers to detect corruption in their book-keeping, though.  malloc's bu debugging interace was removed, otherwise that'd be an alternative to print NULLs when debugging is enabled -- might make sense to ressurrect it, but hopefully not.

(cherry picked from commit 8fb45718daf7200ee0dfa283f3aa38cae757f8a1)