Make some notes about libged+libgcv for ascv5
Hitting an interesting limitation with using libged in a libgcv plugin -
at least on Linux, attempting to do so is crashing when the app trys to
call the libged plugin cleanup on application exit.  Not really sure why
- neither gdb nor the AddressSanitizer can shed much light on it so far
- but it seems to be triggered in some fashion by the dlclose calls for
the libged plugin handles.  Assuming this is an OS limitation, will need
to do a bit of refactoring to put the needed ascv5 logic in librt (which
is arguably where it belongs anyway, at least until libg is stood up...)