First try at supporting cyclic reporting in db_ls
For the qged work, we'll need some way to handle cyclic paths in the
model.  Because a cyclic path object has no non-zero reference counts,
it is not a traditional "tops" object and thus will not be reported by
tops - a comb structure of A u B u A would result in an empty tree view,
despite the fact that there are objects present in such a .g file.

For display purposes, we need to regard cyclic objects as "tops" objects
so we can see them (ideally we will also flag them as problematic
visually, but that's for later.)  Initially, we need a way to detect the
cycles and report them - this experiments with a db_ls flag that alters
the TOPS reporting.