Enable the ASCII v5 libgcv plugin
This isn't ideal (we've not fixed the libged/librt migration issue
manifesting in Archer, and so have to leave the implementation of the
libged command versions of this logic in place) but this setup lets us
complete a no-Tcl build while still generating the sample .g files.
Need to do validation of them against the asc2g variants, but this is a
significant piece of the "core" build functionality that was previously
not working without Tcl.  Wanted to get this stood up, even with work
still to do, so the implementation isn't buried too deeply in the
history.