Don't make fast4-g region names for skipped bots
Customer report of spurious high level groups in fast4-g conversions
using the -c option to constrain the conversion to a particular subset
of the model.  One of the make_region_name calls was getting triggered
even when the region_id was in a skipped range, which was artificially
increasing the maximum active region_id and resulting in empty region
combs being activated for higher level component numbers.

Fix is to guard the bot make_region_name call with a skip_region check.