Merge in changes to support building with Qt6
Take the early work done to test building with Qt6 and make it available
as part of the main build, while also preserving the option to use Qt5.
Qt6 porting isn't quite done (there are a couple places where settings
are just turned off until we find the right Qt6 answer) but this gets us
most of the way there.

Aside from an eventual need to keep up with the current platform, mainly
doing this because Qt6 is somewhat easier for me to build and I want to
see if a debug build can offer insights on what I've done wrong in the
QgModel work to provoke crashes in the main Qt model/view logic.