Work on defining a close command in libged
In principle, we want open, close and man to be libged commands rather
than having to have qged define them - they have sane meanings in the
library context, as long as the parent gedp remains viable.

In theory even "exit" might be implemented as a libged command, if we
have it return the BRLCAD_EXIT flag the way gsh's version currently does
to allow applications to do their own cleanup...  It would be nice if we
can manage it, since these commands are common to essentially all
applications wanting to present a GED command interface.  The
application would have to have logic to check the return flags and
handle the various cases, but eliminating the need for layered command
execution on top of ged_exec would be a significant overall win for
simplicity.