Improve behavior of QgModel/QgTreeView during edit
Somewhat reluctantly, switching to triggering a full model reset after
every ged command that indicates the .g file has changed.  Not sure if
that's going to be problematic for large files, but conceptually it's
fair enough - in principle, any GED command CAN change the whole
database structure.  In order for this to work while keeping the tree
usable, we have to restore the open/close state after the model is reset
(by default it will collapse everything.)  This is a bit tricky, and the
working approach so far appears to be 1) cache the open/close state in
the QgItem when the view sends expansion or collapse signals, 2)
creating a QgTreeView slot that is triggered in response to a db change
signal from the model and 3) iterate over the QgItems in the view slot
comparing their current expansion status to their stored status, and
expanding any that aren't currently expanded but report that they should
be expanded.