rework dynamic library closing, now deferred to avoid racing.
closing dynamically opened libraries on demand was causing a problem
(segfault) on Mac due to our use of C++ constructors/destructors to
load and unload plugins.  when dynamically loading a library that
causes its dependencies to also be dynamically loaded (e.g., loading
libged-dm loads libdm which loads all the dm plugins), there was a
problem when the libraries were closed.  libraries (libdm in this
case) were being unloaded while we were still in that library's
destructor. the plugin that caused that library to be loaded,
libged-dm, was already unloaded so it would crash while still in the
destructor in the process of unloading that library's plugins.

the solution employed here is to defer closing so that the constructor
can complete, and introduce an explicit purge function bu_dlunload()
that actually unloads all the libraries.  this only happens when libbu
is unloaded (again using the destructor method), or bu_dlunload() can
be manually called to force it.