Add some logic for duplicate comb tree entries
Accidentally tested a case I considered early on, but hadn't gotten to
yet - if a comb has two absolutely identical entries (name, matrix, op,
etc.) we have a problem.  (In this case two successive invocations of g
box.r cone.r with moss.g was the trigger.) The gInstance definition,
based on the definition data in the comb, is not unique.  The ONLY thing
that makes the two entries distinct is they have different positions in
the comb tree.

Deal with this by incorporating a counter into the hashing, and then
have the tree walk make sure it's not encountered the current hash in
its walk before we proceed to assign and use it.  This implementation
probably isn't the most efficient possible in that we start over from
zero each time we hit a duplicate rather than stashing the current
highest cnt state for each "0" hash, but hopefully these cases will be
rare enough and small enough that such an optimization won't be needed.

This is effectively the final uniqueness guarantee for anything stored
in a comb tree - if we find anything that would otherwise have been a
duplicate, the counter will keep incrementing until the hash IS unique
within the tree.