Correct db_lookup logic for name with '/' char
Back in 2018, when I added SVN commit r71281 I asserted that it wasn't
valid to have the forward slash '/' in an object name, and consequently
its presence indicated we have a path.  Unfortunately, while that
probably *should* be true, it turns out there exist examples in the wild
that do define their names with forward slashes.  It is possible to
create an example of this in MGED with: make "test.s/" sph

Accordingly, adjust db_lookup to first try a normal lookup on the
provided string without doing anything path related.  If that doesn't
work, we can then go ahead and try and interpret the supplied string as
a path, but if we don't first try the literal string we simply can't get
the correct directory pointer to manipulate in the "test.s/" case.

The presence of such a name does probably indicate either a mistake or
some sort of automated conversion naming, but regardless of how it got
there we have to be able to work with it to fix it.  Since it *can*
exist, correct the librt logic accordingly.