Eliminate the separate libged DM management
Rather than trying to track the display manager pointers separately at
the GED level, we will instead rely on the gvp->dmp pointers.  These can
be updated by client codes without requiring them to know about GED, and
as long as GED keeps track of the views it will also know how to find
all the dmps.  The responsibility of code attaching a dmp to a view is
thus to assign the struct dm pointer to the view's dmp slot, since the
broader codebase won't know about the DM except through its view.