Try to adjust the xxhash logic for Coverity issues
New Coverity scanner doesn't like some of the individual XXH64_update
calls, reporting out of bounds access in the internals.  At a glance
it looks like small objects may be part of the problem?  See if we
can simply feed the structs in whole to the update routine, rater than
calling out the individual slots.  If the updating works correctly this
would be much better anyway, since we could adjust the structs more
easily...

If this either doesn't work for view updating or doesn't clear the
issue, I'll have to dissect more carefully what the internals of the
xxhash routine are wanting as input...