Get Shift-n working for NIRT shooting in qged
This is a rather interesting case, in that (unlike the majority of key
bindings associated with views) running nirt requires much more
extensive knowledge of the application level logic rather than just
manipulating the libqtcad widget's view state.

The two choices are to associate data with the widgets at various levels
so libqtcad has enough info to call ged_exec in a binding, or to have
the application handle the more complex invocation at its own level
rather than in the libqtcad generic bindings.  The latter is a bit
complex, but fortunately most of the machinery is already in place from
the work getting the tree highlighting to respond to mouse events, which
raises similar issues.

We define a key event handler in the application level event filter, and
use Qt's ability to report the currently focused widget and the parent
widgets of that widget to see if our focus is within the central view.

All of this logic (both in libqtcad and qged) should be considered
experimental, since we will also need application-level runtime
configurable mechanisms for defining such bindings.