Add libqtcad default binding toggles to the views
There are two conflicting goals here - one is to have the default
QtCADView widget be as functional as possible for applications, without
requiring a lot of additional work on the application's part to set up
what we think of as basic viewing functionality.  On the other hand, we
want the application to be able to support customizing the bindings - in
which case, low-level baked-in bindings might interfere.  Moreover, the
question of app level bindings can't be fully resolved at the libqtcad
level, since the desired app-level "bound" actions may be more involved
than just altering the view struct and may need more information to
execute than is available at the libqtcad level (ged commands being the
obvious example.)

To support both of these scenarios, we add toggles that will allow the
parent application to opt in or out of the default, pre-baked MGED style
bindings built in to the widgets.  If the application opts out by
disabling them, it is then up to the application itself to implement its
own solution - but this way the widget itself will not do anything
unexpected if the parent app doesn't implement a binding that is defined
by default (after all, that omission in the parent may be intentional.)