Have ert enable the fb in underlay if it is off
When running ert, turn on the display of the embedded framebuffer in
underlay mode if it isn't already enabled.  Without this the visual
output of ert is hidden by default, which is not ideal.  If there's a
use case for running it without enabling the fb by default, we can add
an option to support that.