fix a null-pointer-subtraction warning/error being issued on qsort.
Turns out the OpenBSD implementaiton we're using has optimizations for
sorting by different byte alignments (int and long in addition to
char) and the method employed for determining alignment relied on
offsetting the pointer address from address zero in order to modulo
the element size.  This can cause particular havoc with link time
optimization (LTO) and other strategies that assume strict aliasing.

These optimizations are very similar to other optimized qsort
implementations and came up in a lengthy discussion in the SPEC
benchmark (which has a qsort test based on BSD and lots of other tests
that use that qsort implementation).  Their "fix" was to degrade
everything to char swaps (i.e., eliminate the optimization) which SPEC
rejected on simple principle of the benchmark already being published,
but to my eye it would have negatively affected the benchmark anyways,
so good call...

A better fix I think is the one proposed here where we keep the
optimization, and simply rely on casting the pointer through to a
uintptr_t instead of subtracting null.  Since we're solely dealing
with long/int/char alignments, it should be a safe enough assumption
(at least insofar as being as reliable as subtracting null) and a fix
that will keep LTO optimization happy.

Original error was:

error: performing pointer subtraction with a null pointer has undefined behavior [-Werror,-Wnull-pointer-subtraction]

While a different qsort variant, relevant summary at:
https://www.spec.org/cpu2017/Docs/benchmarks/605.mcf_s.html