Don't link the libs to static libraries
This was prompted by building libqtcad statically, where without the
imported targets being part of the system we didn't get necessary
settings for compilation.  However, really want to be mixing shared and
static libs, and as libqtcad is being build shared-only now the original
motivation for adding it has been removed.