fix windows issue when sources got with git but built with wsl fail.
issue is by default git for windows checks out text files with CRLF
line endings.  this is a problem for awk which is expecting unix-style
LF line endings.

fix used here is to leverage cmake's ability to configure a file and
perform eol conversions, so we copy the awk scripts from source to
binary. this ensures they have LF endings and build logic is updated
to use the build dir version.

as before, we're intentionally avoiding .gitattributes to avoid going
down that road if we don't have to.  will attempt to submit this
upstream.