turn eigen hunting off, just use what we bundle.
header-only libs are not usually worth toggling.  an update to a
system installed header-only lib is (by far) most likely to result in
a compilation failure than it is a useful security or significant
performance improvement.  security issues are already a priority
for all bundled dependencies, significant performance is unlikely.

from a build management perspective, this change will stop the
excessive cmake blather from Third_Party_Headers() calling
find_package() warning that Eigen does not exist as a package.  it was
confusing to some compiling as it looked very much like an error.