Use file structure to assign target folders
Sean figured out that our using the "regress" directory for regression
targets might be causing problems with Visual Studio.  Instead of using
that as a catch-all, assign folders based on the file system directory
in which the targets are defined.  This is largely what we were doing
manually anyway, and also has the advantage of producing a target
organization that lines up with the files on disk.