First cut at making level-appropriate pnts
Give a point, make a stab at putting it through the level transforms to
truncate it to that level - in effect, 'snap' it to the level grid.
Based on the visual output some version of this is probably the heart of
the "decode" operation for POP buffers, but I've not yet dug in to
confirm that - i.e., this may not be correct yet.  Also, need to do
cumulative triangle writing, not individual levels - deeper levels
assume the triangles of higher levels.