this fixes the issue finding the boost libs during compile.
apparently if we just let boost find libs, it ends up finding and using dynamic libs during cmake, but then attempts to link against static libs during linking (which are named differently -- with a lib prefix -- and thus not found).   we either have to specify static-only or dynamic-only to boost in order for the streams to not get crossed.  it works to manually add the boost lib dir to the msvc project after cmake generation, but that doesn't work for msvc 2022's ninja build and is annoying/confusing.  this avoids the issue altogether.