Keep and reuse one GEDP in MGED
See if we can apply a lesson from qged to MGED - keep the gedp once it
is created, and reset it instead of destroying and re-creating the gedp
when a new file is opened.  The latter is more verbose and error prone,
and is not needed since only the dbip and wdbp based data needs to
change.

(cherry picked from commit 288e8933faf99cc64c76becc747a6e0df114612e)