Verify the level is correct at the draw point.
Need to rethink how the LoD is being triggered generally - the CSG
primitives are being handled by draw_update, but that's probably not the
right way to do it - particularly since we won't know how many child
levels we need to iterate.  May be better to have each object set a
callback that knows what to do and is triggered at draw time, instead of
trying to orchestrate at higher levels.