Add an IDN matrix to the lod test
This has no functional implications, but it is being added to help make
it clarify that this is the point at which a hierarchy matrix is applied
to the mesh data. (Its application here is why we don't need to concern
ourselves subsequently with applying matrices to vertices.  It also
means that we aren't able to apply any instancing tricks to use one mesh
+ matrices to reduce our memory footprint without some major design
implications.  However, as far as I can tell the OpenGL versions that
would allow us to do that are newer than our current minimum anyway...)