Start setting up to figure out dlist support
Using OpenGL Display Lists is desirable for drawing performance, but
involves some tradeoffs when it comes to memory size (which is the
reason the MGED gui makes it a user option rather than always enabling
it.)  To generate a dlist we need to hold both the original triangle
data and the OpenGL data in memory until the list generation process is
complete.  At that point the drawing code can clear the original lod
data (we'll need to add a function to support that), but for at least
the duration of the generation process our memory requirements are 2X
that of just directly drawing the triangles.

This extra memory requirement has a number of implications.  Without
splitting up large meshes, there are a category of meshes we simply
won't be able t handle in memory at all.  This is always true as meshes
get arbitrarily large, but the constraint gets tighter if we use dlists.
We need to fall back on NOT using a dlist if the mesh LoD data looks
like it is beyond a threshold size, to avoid the user having to manually
enable/disable dlist usage - they shouldn't have to know anything about
it.

(Part of the LoD goal is to avoid trying to push that much data through
the pipeline in the first place, but for the moment we may still face
this scenario if the user zooms in close to a large mesh.  We would need
the hierarchical breakdown of the mesh and selective loading to cope
with that situation - we may get there, but don't have that capability
yet - early tests were disabled until we get the initial capability
working.)

There will certainly be more implications, but one definite requirement
will be that the loading code must recognize when data has been
completely cleared to save memory and a full reload is required in
situations that would otherwise involve only a partial load.  There is a
tradeoff here, since a full load involves reading in all the lower level
files that can otherwise be kept in memory (avoids file I/O overheads.)
It may be that for sufficiently small in-memory footprints it would be
worthwhile to keep the non-dlist data and pay the memory overhead, and
only start purging when memory costs get high - this would save a lot of
small file accesses for "zoomed out" items that only require small
meshes.  Accordingly, the loading code will simply look at the current
state of the points list to see if prior data is found in memory or not.

We will need bu_mem to asssess the current memory state of the system
and what our available resources are - the gl_draw_tri code will be the
point at which we make the decision to create a dlist or proceed with
the "normal" drawing, so we'll also need a way to report the memory size
of the lod data in user code.