Need two-sided shading for LoD drawing
When we have thin objects (like a windshield) or we get to very low
levels of detail, the snapping can result in collapsed triangles - i.e.
both sides of the solid drawing a triangle in the same place.  That
means if we're providing normals, we can end up with the "wrong"
triangle being the last one drawn for a given view, depending on the
ordering.   At low detail levels this effect manifests as "speckling"
and is a bit annoying, but it is extremely apparent on large thin
solids. OpenGL has a lighting model for this situation, so we enable it
to avoid jarring visual artifacts.