Apply the matrix at draw time, not in initial load
Rather than loading the mesh data with the matrix applied, cache it once
and apply the matrix from the scene object's information at draw time.

This saves on cache calculations for models using matrices, While in
principle this would allow us to store one lod structure per unique mesh
and reuse that for multiple scene objects, that is only viable in
orthogonal drawing - in perspective mode, we will most likely want to
load different LoD levels in different instances of the same object as a
function of distance from the camera, which would make such reuse
untenable.

For xpushed models this approach doesn't buy us anything even in
orthogonal mode since we don't have non-unique meshes anyway, but on
models making heavy use of matrices to instance objects the processing
and disk space savings are significant.

Another benefit (and the reason this is being done now) is we can look
up a stored cache key based on the directory pointer of the solid,
without having to try and incorporate the matrix hash into the look-up.
The latter would be somewhat unstable, since we calculate the matrix on
the fly during tree walk.  Floating point fuzz would most likely result
in the same walk not always producing exactly the same matrix, and thus
not hashing to exactly the same value (npush has to cope with similar
issues, but it retains the full matrix to do so.)  That would make
reliable cache lookups iffy and shouldn't be needed in any case, if we
can simply defer the matrix handling until the drawing stage.