First cut at visibility-based LoD updates
When we zoom in on a large model, there's no need to update data for
objects we're not going to see.  Generate an oriented bounding box from
the view info and use the SAT test to identify which objects need to
have their LoD adjusted.

This just leaves the previous levels as-is - may be worth experimenting
with reducing the memory footprint of non-visible objects and/or
skipping the drawing operation for objects we can flag as not visible.