Enable C++14 (will eventually be needed for openNURBS)
Was hitting an error in damlevlim.cpp on Windows having to do with
assert failures.  It turns out that when this file was adapted from the
C++17 upstream version, I missed a C++14 requirement implicit in the
std::mismatch function.  There IS an older version of that function, but
it is unsafe when the second string is shorter (which of course happens
frequently with this function.)  Rather irritatingly, Linux apparently
compiles and runs this fine even with the C++11 flags set, so it was
only when the Windows runtime failure triggered that it was caught.

This function can be re-implemented using something like vls, but since
we're going to need C++14 at some point for the more modern openNURBS
anyway, go ahead and bump the standard now rather than doing an
ephemeral rework.