Stash a test of double to/from long int conversion
When "snapping" vertices from double or float to integer types for
robustness, we want (in principle) as little information loss as we can
manage.  This test is experimenting with how to determine what those
limits are dynamically based on input information.

Did this as part of some early LoD experiments - ended up going another
direction for that work, but this idea may still be needed/useful down
the road so stash it for reference.