Note some problems with CSG LoD that need work
With the LMDB refinement, triangle LoD is now working reasonably well in
a couple of tests.  The next obvious problem is in mixed bot/csg models
where rotating around is growing memory alarmingly.

This commit fixes an obvious problem that was triggering a redraw on
every view change regardless of whether it was needed, but it doesn't
address the underlying issues that this mistake exposed.  The reuse of
vlists, which is a staple of vlist drawing for performance, doesn't
scale when we start doing lots of very high detail vlists for many many
zoomed in solids.  We need to do view culling the way we do for BoTs to
limit hyper zooming to only visible primitives, and look at how we're
handling vlist management for these wireframes in LoD mode.

If we have to we can allow the user to enable/disable LoD for bots and
csg selectively, but they really shouldn't have to do that in the first
place.  I don't want to complicate the usability unless it proves much
more difficult than I expect to make the necessary improvements to CSG
handling so everything can Just Work...