First stab at constructing the frustum
Still working through how to construct what we need for intersection
testing.  By default, it looks like the eye point is inside the scene,
which means the near plan is also in the scene and the frustum doesn't
actually bound what's drawn.  We might want near distance == 0 to get
the full pyramid, but need to check if the SAT test will tolerate that.