Start adding completion back in to QtConsole
I had originally removed the completion support from the Paraview
console widget while integrating it for simplicity.  However, now that
we're getting to the point where tab completion is a question, we're
probably going to want it.

Stubbing in the methods and calls - untested, as we'll need a fair bit
of machinery in place before we can kick these particular tires...