Make gl_draw_tri into a generic bv_scene_obj call.
The mesh LoD data is only one of a number of possible specific types we
may want to handle - rather than tie up a callback slot with each
specific type, just pass the scene obj through and use the type flags
"under the hood" to figure out which specific drawing routines to use
for specific scene objects.

This should allow for more flexibility and much less code changing as we
experiment with different ways to draw scene objects.