Move the definition of a torus to libbg
Wanting to do some experiments with polyline LoD plotting using torus as
a test subject.  There's nothing inherently specific to librt about the
structure itself, so move it down to libbg to make the concept of a
geometric torus available at that level.

This shouldn't have any implications for librt, as from its perspective
this is just a structure/magic name change (s expressions have been
added to CHANGES file.)