Various adjustments to drawing LoD test.
Behaving somewhat better now - still not an enormous speed-up, but looks
like maybe a bit of an improvement, and if we ever switch to more modern
APIs we might be able to do better.  More importantly, this approach
offers the possibility of reclaiming transient large memory usages.

Fundamentally, the CSG LoD was intended to increase drawing detail
rather than decreasing it, and its settings are configured
accordingly.