Allow separate toggling of mesh and csg LoD
Since we're usually wanting mesh LoD to decrease drawn information and
CSG LoD to increase it, we may want to enable/disable them separately
depending on whether we're trying for better visuals or to reduce memory
footprint.  "view lod" will still control them in sync, but allow for
individual toggling as well.