Print the prefix if a deep path is given to search
When search is given a deep path (for example, /all.g/box.r) it will do
the search using box.r as the starting point.  Previously, this meant
that the results would be printed using box.r as the root:

/box.r
/box.r/box.s

This may be surprising when the full path is supplied, particularly if
the caller wants the results to be relative to the full specified path.
The adjustment here keeps the prefix when reporting the search results:

/all.g/box.r
/all.g/box.r/box.s

Note that under the hood the search is still conducted relative to
box.r.  In particular, a -depth 1 option will start counting from box.r,
not all.g - this is a bit unintuitive, but a fix will require somewhat
deeper changes to the search logic to recognize the initial depth of the
supplied path and process accordingly.