this gets art working with the perspective camera too!
finally figured out the discrepancy.  rt computes a view-rot-scale
matrix in order to calculate the eye and view2model view matrix.
appleseed however needs a view-rot matrix as it calculates the scale
based on camera parameters.  result was the view not matching because
of a double application of viewsize (which in a perspective view
results in massive zooming).

this recalculates the view2model matrix without viewsize incorporated
and it works like a charm.