match rt's perspective mode rendering even better.
perspective rendering was a half viewsize off and it's not yet exactly clear to me why, but this makes it match exactly for at least standard az/el view descriptions.  suspect there's a factor inside rt's perspective mode view setup (that probably shouldn't be there).  regardless, we reinstate the viewsize*0.5 so orthogonal and perspective can both call do_ae() as-is.  no hints looking back through time on why the 0.5 is in there.