Consolidate the transparency variables
Rather confusingly, the opengl display managers (the only ones that
currently support transparency) use both an internal variable and an
entry in the dm calltable to track and report transparency.

This change consolidates to use the internal variable, with a
getTransparency function replacing the dm_transparency variable.  Any
backend that implements a transparency setting can define it the same way
the OpenGL backends define their mvars settings.

Longer term we may also want to replace all the variable specific
function pointers with some variation on _dm_cmd_get so we don't
have to keep changing the calltable every time a backend wants to expose
a new setting...