add geometry visibility script!
This script calculates visibility for specified eye points at given base object(s) with optional occluder(s).  It autogenerates a stack of images, labels, and more into an annotated top-down and perspective view rendering for those eye points with exceptional analytic effect.  this script really could be turned into a full-fledged generally-useful application, but was rapidly prototyped for a real use case.  it's included here in a raw state to preserve it.