merge system path searching from bu_whereis into bu_which
while they are different with bu_whereis specifically searching system paths and bu_which only looking in the user's PATH, that's not a helpful distinction when we're looking for programs.  moreover, our whereis behavior wasn't strictly conforming as system whereis also reports back all matching results and we only returned the first.  so this change has it look in both user and system locations until it finds the command in question.  probably should rename as behavior is no longer posix-striving.