remove the USE_FORKED_THREADS hack.
it was implemented in support of the 2001 linux kernel (2.4) which had a bug affecting thread accounting.  this workaround spawned all threads in a subprocess, whose accounting did not have the bug so rt statistics could be calculated accurately.  support for the 2.4 kernel ended in 2011, so plenty of time to let the hack go.  if it's still needed, it may make more sense to have libbu simply do its own time accounting.