try using FILE(STRINGS ... REGEX ...)
Either the behavior of file() is not consistent across versions of cmake when splitting across lines, or the behavior of cmakelists() is inconsistent in how it deals with mergeing lists with spaces.  This change checks the prior to see which is problematic in order to understand where the limitation resides.