Reconnect the fb_mode toggle
This button exposes a weakness in the current signal/slot design for
view management - the button will toggle the view, but does not update
the view settings plugin.  This is a side effect of needing to avoid
triggering an event which will force GUI updates if the gui is itself
the origin of the action to initiate change, which has potential for
infinite looping, but in this case also results in one GUI element not
correctly reflecting a view state change as a consequence of action by
another GUI widget.  We also must bear in mind that both elements may
need updating as a consequence of view state changes driven by a command
line command.

Need to take a step back and consider what changes are needed.