Start thinking about object set selection.
For editing, view drawing object changes with graphical selection, etc.
we want to hone in on the concept of "working sets" of objects.  We
don't have sed/oed, but we do have to have some sense of which object or
objects in the who set we intend to apply edit operations to.  We also
want Archer's ability to graphically identify objects to erase from the
current drawn set, and probably other abilities as well.