Not using hash mode - remove.
Coming up against a long standing problem - commands like draw and erase
don't have a good way to specify specific instances of comb trees.  For
example, if we have:

a
  u [M1] b
  u [M2] b
  - [M3] b

and we want to draw or erase just the M2 instance of a/b, we don't have
a way to specify it.  This is a problem when operating the graphical
erase plugin on scene object paths, since the scene objects correspond
to specific a/b style instances but we don't have any way to translate
those concepts into individual GED command manipulations.  (Nor is the
problem unique to that case - it is just the most current example.)  The
bookkeeping needed for drawing contraindicates trying to evolve some
sort of special case manipulation for this particular situation - we
need a way for GED commands (and librt prep, for that matter) to
individually specify specific comb tree instances.  The same problem
arises when we think about defining custom sets of instances based on
selections from the scene.