Checkpoint work on hierarchy instance specification
Got as far as draw and erase in qged being able to draw and erase
specific comb instances using the "@#" notation for paths.

However, ert (embedded raytracing) does not work, because the raytracing
layer doesn't correctly process the new path syntax.  libged can pass
the paths to rt, but rt won't know what to do with them.  Adding support
for the new syntax will involve some scary changes to the core of the
librt tree walking layers, so it's time for a branch.

This is painful, but appears to be necessary for proper support of a
number of graphical interface features.  It first came up when using the
erase mouse tool to clear objects out of a scene - reported paths in
"who" weren't unique after such erasures, highlighting the lack of a way
to address specific instances on the command line.  Such interaction
modes make it more likely we'll need to raytrace otherwise atypical
lists of instances, which will eventually result in an incorrect visual
as the user erases one instance and rt interprets the specifier as a
different instance.  Another, more readily observable problem is what
happens if we want to add support to the tree widget for selecting and
drawing specific instances in the tree - the tree exposes all instances,
but to visualize them we need to pass a unique string specifier to the
draw command, and we don't (or didn't, at any rate) have a way to do so.
And even if we were to add some convoluted hack specific to the tree to
try and work around this, we'd still have the ert problem.

Once it is clear what logic changes are needed for this to work
correctly, we can decide how to introduce it back into main.  There's a
decent chance we can't use the environment variable guards used for the
GED commands for speed reasons, so we may have to introduce a new
function stack with the intent of it eventually replacing the current
logic once it proves out.  For testing purposes in the branch we'll make
the changes in-situ to avoid unnecessary complications, but it would
almost certainly be too risky to merge the changes in that form.