Test adding a behavior flag to the db_i struct
Since the comb instance path support is disruptive, we need a way to
enable/disable it (at least in the early stages) so we don't break
existing logic.  dbip is present when we are setting up db_full_path
structures, so that is where we need to add a setting if we aren't going
to otherwise change the API.  As long as the fp_cinst array is zero
(i.e. its default) the strings generated for full paths will be the
standard strings of current behavior, and if we can key the setting of
fp_cinst based on this value being 1, we should be able to allow qged to
enable this behavior while leaving it off for everyone else by default.