The previous commit is not going to work. Object full paths need to be preserved when checking to ensure hierarcial information is maintainted.
Major change to find_cmd_line_obj so that now the full path is compared by split along slashes instead of only checking the parent. Also changed the return type to be the object pointer instead of an int for the index for a more intuitive return.
Added error checking and a log cap to make the issue apparent but reduce memory use and unnecessary blather.